Add first attempt at LaTeX => SVG
This commit includes a basic ugly Node.js script that runs latex and
dvisvgm on some sample LaTeX code. It then inserts the SVG code inside a
string of HTML and writes that to a file.

I'm basically committing this to save progress, since I'm about to
refactor things so that the code is closer to its Gatsby plugin form.